+++ src/seed.ts
@@ -1,8 +1,8 @@
 // Imports data from Data.json
-import data from "@/lib/resources/data/data.json";
+import data from "@/backend/resources/data/data.json";
 
 // Imports Sports Schema
-import Sport from "@/lib/resources/models/Sport";
+import Sport from "@/backend/resources/models/Sport";
 
 // Function that creates many sports model in the database
 export default async function seedAll() {
